Remove README.md
We cannot set sparse checkout early for config init since we do not have any
hooks to run. So it may conflict with any README.md file in the home dir and
init may fail in that case unless we move or remove it.

Better put README on another branch if needed.